study-agents-differences/pyproject.toml
[project]
name = "study-agents-differences"
version = "0.2.0"
description = "Comprehensive comparison study of AI agent frameworks"
requires-python = ">=3.12"
dependencies = [
    # --- Core ---
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
    "tiktoken>=0.9.0",
    "numpy>=2.2.3",
    "httpx>=0.28.1",
    "tavily-python>=0.5.1",
    "streamlit>=1.42.1",

    # --- Agno ---
    "agno>=2.5.0",

    # --- LangGraph ---
    "langgraph>=1.1.3",
    "langgraph-checkpoint>=2.0.16",
    "langchain-core>=0.3.37",
    "langchain-openai>=0.3.6",
    "langchain-huggingface>=0.1.2",
    "langchain-community>=0.3.0",

    # --- LangChain ---
    "langchain>=1.0.0",

    # --- LlamaIndex ---
    "llama-index>=0.12.19",
    "llama-index-core>=0.12.19",
    "llama-index-agent-openai>=0.4.6",
    "llama-index-llms-openai>=0.3.20",
    "llama-index-llms-azure-openai>=0.3.0",
    "llama-index-llms-huggingface-api>=0.3.1",
    "llama-index-embeddings-openai>=0.3.1",
    "llama-index-readers-file>=0.4.5",

    # --- OpenAI (raw API agent) ---
    "openai>=1.63.2",

    # --- OpenAI Agents SDK ---
    "openai-agents>=0.12.5",

    # --- Pydantic AI ---
    "pydantic-ai>=1.70.0",

    # --- CrewAI ---
    "crewai[tools]>=1.12.2",

    # --- Google ADK ---
    "google-adk>=1.26.0",
    "litellm>=1.61.11",

    # --- Smolagents ---
    "smolagents[toolkit]>=1.24.0",

    # --- Strands Agents SDK ---
    "strands-agents>=1.32.0",

    # --- Claude Agent SDK ---
    "claude-agent-sdk>=0.1.52",

    # --- AG2 ---
    "ag2[openai]>=0.11.4",

    # --- Microsoft Agent Framework ---
    "agent-framework>=1.0.0rc5",

    # --- Embeddings & Vector ---
    "chromadb>=0.4.0",
    "sentence-transformers>=3.4.1",
    "fastembed>=0.3.0",

    # --- Azure ---
    "azure-identity>=1.20.0",

    # --- Utilities ---
    "tabulate>=0.9.0",
    "rich>=13.9.4",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
    "ipykernel>=6.29.5",
]
